+++ x/btcstaking/types/msg.go
@@ -181,6 +181,8 @@ func (m *MsgBtcStakeExpand) GetStakeExpansion() (*ParsedCreateDelStkExp, error) 
 	return &ParsedCreateDelStkExp{
 		PreviousActiveStkTxHash: previousActiveStkTxHash,
 		OtherFundingOutput:      otherOutput,
+		FundingTxHash:           fundingTxHash,
+		FundingOutputIndex:      idxOtherInput,
 	}, nil
 }